RFCT: move image manager to respective folder for better maintainability
